refactor: remove insertbreak helpers :recycle:
