--- packages/core/src/helpers/insertBreak/index.tsx
@@ -1,77 +1,0 @@
-import { Text, Element } from 'slate';
-import { NodeFunction } from '../../config';
-
-/**
- * Inherit the type of the current Element when you press enter.
- */
-export const copyInsertBreak: NodeFunction = (editor, entry) => {
-  const [currentNode] = entry;
-  if (!Element.isElement(currentNode)) {
-    editor.splitNodes({ always: true });
-    return true;
-  }
-  editor.splitNodes({ always: true });
-  return true;
-};
-
-/**
- * Inherit the type of the current Element when you press enter.
- * If the text is empty, remove the Element.
- */
-export const copyAndRemoveEmptyTextInsertBreak: NodeFunction = (
-  editor,
-  entry
-) => {
-  const [currentNode, currentPath] = entry;
-  if (!Element.isElement(currentNode)) {
-    editor.splitNodes({ always: true });
-    return true;
-  }
-  const textList = Array.from(
-    editor.nodes({
-      at: currentPath,
-      match: (node) => Text.isText(node),
-    })
-  ).map((node) => (Text.isText(node[0]) ? node[0].text : ''));
-  const isTextEmpty = !textList.join();
-
-  if (isTextEmpty) {
-    const next = editor.next();
-    if (!next) {
-      return false;
-    }
-    const [nextNode, nextPath] = next;
-    const parentEntry = editor.parent(nextPath);
-    const textNodeList = Array.from(
-      editor.nodes({
-        at: parentEntry[1],
-        match: (node) => Text.isText(node),
-      })
-    );
-    const [lastNode, lastNodePath] = textNodeList[textNodeList.length - 1];
-    if (!Text.isText(lastNode)) {
-      return true;
-    }
-    if (Element.isElement(nextNode)) {
-      if (editor.isVoid(nextNode)) {
-        editor.removeNodes({ at: currentPath });
-        return true;
-      }
-    }
-    editor.setSelection({
-      anchor: {
-        path: lastNodePath,
-        offset: lastNode.text.length,
-      },
-      focus: {
-        path: lastNodePath,
-        offset: lastNode.text.length,
-      },
-    });
-    editor.removeNodes({ at: currentPath });
-    return true;
-  }
-  editor.splitNodes({ always: true });
-
-  return true;
-};
