+++ packages/element-list/src/item/index.ts
@@ -7,11 +7,14 @@ const element: Element<Attributes> = {
   type: 'list-item',
   attributes,
   editable,
-  insertBreak: (editor, nodeEntry, config) =>
-    helpers.insertBreaks.copyAndRemoveEmptyTextInsertBreak(
-      editor,
-      nodeEntry,
-      config
-    ),
+  insertBreak: (editor, nodeEntry) => {
+    const [node] = nodeEntry;
+    if (helpers.textHelpers.isTextEmpty(node)) {
+      helpers.nodeHelpers.removeNode(editor, nodeEntry);
+    } else {
+      editor.splitNodes({ always: true });
+    }
+    return true;
+  },
 };
 export default element;
