+++ packages/element-list/src/ordered/index.ts
@@ -32,12 +32,15 @@ const element: Element<Attributes> = {
     }
     return isNormalized;
   },
-  insertBreak: (editor, nodeEntry, config) =>
-    helpers.insertBreaks.copyAndRemoveEmptyTextInsertBreak(
-      editor,
-      nodeEntry,
-      config
-    ),
+  insertBreak: (editor, nodeEntry) => {
+    const [node] = nodeEntry;
+    if (helpers.textHelpers.isTextEmpty(node)) {
+      helpers.nodeHelpers.removeNode(editor, nodeEntry);
+    } else {
+      editor.splitNodes({ always: true });
+    }
+    return true;
+  },
 };
 
 export default element;
