+++ packages/element-toggle/src/body/index.ts
@@ -1,4 +1,4 @@
-import { Element, helpers } from '@bento-editor/core';
+import { Element } from '@bento-editor/core';
 import attributes, { Attributes } from './attributes';
 import editable from './editable';
 export * from './renderer';
@@ -7,7 +7,9 @@ const element: Element<Attributes> = {
   type: 'toggle-body',
   attributes,
   editable,
-  insertBreak: (editor, nodeEntry, config) =>
-    helpers.insertBreaks.copyInsertBreak(editor, nodeEntry, config),
+  insertBreak: (editor) => {
+    editor.splitNodes({ always: true });
+    return true;
+  },
 };
 export default element;
